+++ packages/devtools_app/lib/src/shared/analytics/analytics_common.dart
@@ -29,8 +29,8 @@ import '../primitives/utils.dart';
 /// }
 /// ```
 ///
-/// Then, add your fields to the [GtagEventDevTools] factory constructor and add
-/// a corresponding getter in the class.
+/// Then, add your fields to the [DevToolsAnalyticsEvent] factory constructor
+/// and add a corresponding getter in the class.
 abstract class ScreenAnalyticsMetrics {}
 
 /// The character limit for each event parameter value sent to GA4.