+++ packages/devtools_shared/CHANGELOG.md
@@ -3,6 +3,12 @@ Copyright 2025 The Flutter Authors
 Use of this source code is governed by a BSD-style license that can be
 found in the LICENSE file or at https://developers.google.com/open-source/licenses/bsd.
 -->
+# 13.0.0-wip
+* Deprecated `DevToolsStoreKeys.analyticsEnabled` since this is only used for legacy analytics.
+* **Breaking change:** Removed legacy analytics APIs and state cleanup (e.g. `apiGetFlutterGAEnabled`, `apiGetDevToolsEnabled`, `apiSetDevToolsEnabled`).
+* **Breaking change:** Removed public constant `devToolsEnabledPropertyName`.
+* **Breaking change:** Removed the `analyticsEnabled` getter and setter from `DevToolsUsage`.
+
 # 12.1.0
 * Adds additional logging to `IntegrationTestRunner`.
 * Fixes assertion error when parsing Flutter channel versions using the format `X.XX.X-X.X.pre-XXX`.